Prova 2/requirements.txt
flask
wtforms
newsapi-python
flask-wtf
Image
flask-sqlalchemy